+++ src/routes/AvailableRoutes.tsx
@@ -1,0 +1,20 @@
+import TokenStore from "../store/TokenStore";
+import DefaultRoutes from "./DefaultRoutes";
+import NotAuthenticatedRoutes from "./NotAuthenticatedRoutes";
+import TechnicalWorkRoutes from "./TechnicalWorkRoutes";
+
+function AvailableRoutes() {
+    const { isAuth, isTechnicalWork } = TokenStore;
+
+    if (!isAuth) {
+        return NotAuthenticatedRoutes();
+    }
+
+    if (isTechnicalWork) {
+        return TechnicalWorkRoutes();
+    }
+
+    return DefaultRoutes();
+}
+
+export default AvailableRoutes;
